+++ apps/commerce-collector/src/main/java/com/loopers/domain/eventLog/EventLogService.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.eventLog;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
