+++ apps/commerce-collector/src/main/java/com/loopers/domain/handled/EventHandledRepository.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.handled;
 
 public interface EventHandledRepository {
