+++ apps/commerce-collector/src/main/java/com/loopers/domain/handled/EventHandledService.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.handled;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
