+++ apps/commerce-collector/src/main/java/com/loopers/domain/metrics/ProductMetrics.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.metrics;
 
 import jakarta.persistence.*;
 import lombok.*;
