+++ apps/commerce-collector/src/main/java/com/loopers/domain/metrics/ProductMetricsRepository.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.metrics;
 
 public interface ProductMetricsRepository {
