+++ apps/commerce-collector/src/main/java/com/loopers/domain/metrics/ProductMetricsService.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.metrics;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
