+++ apps/commerce-collector/src/main/java/com/loopers/infrastructure/EventHandledJpaRepository.java
@@ -1,6 +1,6 @@
 package com.loopers.infrastructure;
 
-import com.loopers.domain.EventHandled;
+import com.loopers.domain.handled.EventHandled;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
