+++ apps/commerce-collector/src/main/java/com/loopers/interfaces/consumer/CatalogEventConsumer.java
@@ -1,7 +1,7 @@
 package com.loopers.interfaces.consumer;
 
-import com.loopers.domain.ProductMetricsCount;
-import com.loopers.domain.ProductMetricsService;
+import com.loopers.domain.metrics.ProductMetricsCount;
+import com.loopers.domain.metrics.ProductMetricsService;
 import com.loopers.kafka.config.KafkaConfig;
 import com.loopers.kafka.dto.KafkaMessage;
 import lombok.RequiredArgsConstructor;
